# docs/swagger.json
{
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a REST API server for threat registry management.\nIt provides endpoints for managing products, instances, threats, controls, domains, and tags.",
        "title": "Threatreg API",
        "contact": {
            "name": "API Support",
            "email": "support@threatreg.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/controls": {
            "get": {
                "description": "Get a list of all security controls in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Controls"
                ],
                "summary": "List all controls",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Control"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new security control with the provided title and description",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Controls"
                ],
                "summary": "Create a new control",
                "parameters": [
                    {
                        "description": "Control creation request",
                        "name": "control",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateControlRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Control"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/controls/{id}": {
            "get": {
                "description": "Get a specific security control by its UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Controls"
                ],
                "summary": "Get a control by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Control ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Control"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a control's title and/or description",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Controls"
                ],
                "summary": "Update a control",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Control ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Control update request",
                        "name": "control",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateControlRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Control"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a security control by its UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Controls"
                ],
                "summary": "Delete a control",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Control ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.CreateControlRequest": {
            "type": "object",
            "required": [
                "description",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handlers.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "handlers.UpdateControlRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "models.Control": {
            "type": "object",
            "properties": {
                "controlAssignments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ControlAssignment"
                    }
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "threatControls": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ThreatControl"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.ControlAssignment": {
            "type": "object",
            "properties": {
                "control": {
                    "$ref": "#/definitions/models.Control"
                },
                "controlID": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "threatAssignment": {
                    "$ref": "#/definitions/models.ThreatAssignment"
                },
                "threatAssignmentID": {
                    "type": "integer"
                }
            }
        },
        "models.Domain": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "instances": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Instance"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Instance": {
            "type": "object",
            "properties": {
                "domains": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Domain"
                    }
                },
                "id": {
                    "type": "string"
                },
                "instanceOf": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "product": {
                    "$ref": "#/definitions/models.Product"
                },
                "threatAssignments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ThreatAssignment"
                    }
                }
            }
        },
...
        "models.Product": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "instances": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Instance"
                    }
                },
                "name": {
                    "type": "string"
                },
                "threatAssignments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ThreatAssignment"
                    }
                }
            }
        },
...
        "models.Threat": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "threatAssignments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ThreatAssignment"
                    }
                },
                "threatControls": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ThreatControl"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.ThreatAssignment": {
            "type": "object",
            "properties": {
                "controlAssignments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ControlAssignment"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "instance": {
                    "$ref": "#/definitions/models.Instance"
                },
                "instanceID": {
                    "type": "string"
                },
                "product": {
                    "$ref": "#/definitions/models.Product"
                },
                "productID": {
                    "type": "string"
                },
                "threat": {
                    "$ref": "#/definitions/models.Threat"
                },
                "threatID": {
                    "type": "string"
                }
            }
        },
        "models.ThreatControl": {
            "type": "object",
            "properties": {
                "control": {
                    "$ref": "#/definitions/models.Control"
                },
                "controlID": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "threat": {
                    "$ref": "#/definitions/models.Threat"
                },
                "threatID": {
                    "type": "string"
                }
            }
        },
...
    },
...
}